requirements.txt
# ------------------------------------------------------------------------------
# requirements.txt
# Minimal Python deps to run Cypher scripts and .env configs.
# ------------------------------------------------------------------------------
neo4j>=5.22.0
python-dotenv>=1.0.1
